requirements.txt
streamlit
pandas
numpy
tensorflow-cpu
joblib
scikit-learn
watchdog
matplotlib